Refactor MainNavigation and remove AuthButton
