+++ app/actions.tsx
@@ -1,0 +1,14 @@
+'use server'
+
+import { cookies } from 'next/headers';
+import { createClient } from '@/lib/supabase/server';
+import { redirect } from 'next/navigation';
+
+export const signOut = async () => {
+  const cookieStore = cookies();
+  const supabase = createClient(cookieStore);
+
+  await supabase.auth.signOut();
+
+  return redirect('/login');
+};
