+++ components/MainNavigation.tsx
@@ -1,9 +1,15 @@
+import { cookies } from 'next/headers';
+import { createClient } from '@/lib/supabase/server';
+import { signOut } from '@/app/actions';
 import Link from 'next/link';
 import Logo from '@/components/Logo';
 import ThemeToggleButton from '@/components/ThemeToggleButton';
-import AuthButton from '@/components/AuthButton';
 
-export default function MainNavigation() {
+export default async function MainNavigation() {
+  const cookieStore = cookies();
+  const supabase = createClient(cookieStore);
+  const { data: { user } } = await supabase.auth.getUser();
+
   return (
     <header className="w-full flex flex-col items-center fixed top-0 z-40">
       <nav
@@ -20,7 +26,23 @@ export default function MainNavigation() {
               </Link>
             </li>
             <li>
-              <AuthButton />
+              {user ? (
+                <div className="flex items-center gap-4">
+                  Hi, {user.email}!
+                  <form action={signOut}>
+                    <button className="py-2 px-4 rounded-md no-underline bg-btn-background hover:bg-btn-background-hover">
+                      Logout
+                    </button>
+                  </form>
+                </div>
+              ) : (
+                <Link
+                  href="/login"
+                  className="py-2 px-3 flex rounded-md no-underline bg-btn-background hover:bg-btn-background-hover"
+                >
+                  Login
+                </Link>
+              )}
             </li>
             <li>
               <ThemeToggleButton />
